requirements.txt
requests
python-dotenv
packaging
